refactor(command-palette): extrait HeaderNav, label raccourci multi-plateforme, reset état à la fermeture
- RaccourciKbd : label ⌘K (Apple) / Ctrl+K (Windows/Linux) via détection
  plateforme, source unique SSR-safe (useSyncExternalStore)
- HeaderNav : regroupe trigger recherche + palette + menu utilisateur, allège __root
- close() réinitialise l'état interne (activeItem/query) : plus de page d'actions
  résiduelle à la réouverture
- aligne les commentaires sur le comportement client (plus de recherche serveur)
