+++ apps/kpilote-webapp/src/routes/__root.tsx
@@ -1,19 +1,8 @@
 import type { QueryClient } from '@tanstack/react-query'
-import {
-  createRootRouteWithContext,
-  Link,
-  Outlet,
-  useLocation,
-  useNavigate,
-  useSearch,
-} from '@tanstack/react-router'
+import { createRootRouteWithContext, Link, Outlet } from '@tanstack/react-router'
 import { TanStackRouterDevtools } from '@tanstack/react-router-devtools'
-import { Search } from 'lucide-react'
-import { useState, type ReactNode } from 'react'
 
-import { CommandPalette } from '@/components/command-palette/CommandPalette'
-import { UserMenu } from '@/components/UserMenu'
-import { Button } from '@/components/ui/Button'
+import { HeaderNav } from '@/components/HeaderNav'
 import { Marianne } from '@/components/ui/Marianne'
 import type { Auth } from '@/auth'
 
@@ -28,8 +17,6 @@ export const Route = createRootRouteWithContext<RouterContext>()({
 
 function RootComponent() {
   const { auth } = Route.useRouteContext()
-  const navigate = useNavigate()
-  const [paletteOpen, setPaletteOpen] = useState(false)
 
   return (
     <div className="flex min-h-screen flex-col bg-background text-text">
@@ -52,39 +39,7 @@ function RootComponent() {
             </span>
           </Link>
 
-          <nav className="flex items-center gap-1 sm:gap-2">
-            {auth.isAuthenticated ? (
-              <button
-                type="button"
-                onClick={() => setPaletteOpen(true)}
-                aria-label="Ouvrir la recherche"
-                className="flex items-center gap-2 rounded-md border border-border bg-surface px-2 py-1.5 text-sm text-text-subtle transition-colors hover:border-border-strong hover:text-text sm:px-3"
-              >
-                <Search className="size-4" />
-                <span className="hidden sm:inline">Rechercher…</span>
-                <kbd className="ml-4 hidden rounded border border-border bg-surface-tinted px-1.5 py-0.5 font-mono text-[10px] font-medium text-text-muted sm:inline-block">
-                  ⌘K
-                </kbd>
-              </button>
-            ) : null}
-
-            <NavLink>Tableau de bord</NavLink>
-
-            <div className="mx-2 hidden h-6 w-px bg-border sm:block" />
-
-            {auth.isAuthenticated && auth.user ? (
-              <UserMenu
-                user={auth.user}
-                onLogout={() => {
-                  void auth.logout()
-                }}
-              />
-            ) : (
-              <Button size="sm" type="button" onClick={() => void navigate({ to: '/login' })}>
-                Se connecter
-              </Button>
-            )}
-          </nav>
+          <HeaderNav auth={auth} />
         </div>
       </header>
 
@@ -103,27 +58,7 @@ function RootComponent() {
         </div>
       </footer>
 
-      {auth.isAuthenticated ? (
-        <CommandPalette open={paletteOpen} onOpenChange={setPaletteOpen} />
-      ) : null}
-
       {import.meta.env.DEV && <TanStackRouterDevtools position="bottom-right" />}
     </div>
-  )
-}
-
-function NavLink({ children }: { children: ReactNode }) {
-  const search = useSearch({ strict: false })
-  const { pathname } = useLocation()
-  const isActive = pathname.startsWith('/indicateurs') || pathname.startsWith('/paniers')
-  return (
-    <Link
-      to="/indicateurs"
-      search={{ individu: search.individu, referentiel: search.referentiel }}
-      data-status={isActive ? 'active' : undefined}
-      className="rounded-md px-3 py-2 text-sm font-medium text-text-muted transition-colors hover:bg-surface-tinted hover:text-primary data-[status=active]:bg-primary-tinted data-[status=active]:text-primary"
-    >
-      {children}
-    </Link>
   )
 }
